# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ViP Dante VM Console API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type LineState = typeof LineState[keyof typeof LineState];


export const LineState = {
  idle: 'idle',
  standby: 'standby',
  onair: 'onair',
} as const;

export interface Line {
  id: string;
  state: LineState;
  label: string;
  updatedAt: string;
}

export type LineInputState = typeof LineInputState[keyof typeof LineInputState];


export const LineInputState = {
  idle: 'idle',
  standby: 'standby',
  onair: 'onair',
} as const;

export interface LineInput {
  state?: LineInputState;
  /** @maxLength 120 */
  label?: string;
}

export interface ResetResult {
  ok: boolean;
}

export interface Vm {
  id: number;
  name: string;
  url: string;
  phoneNumber: string;
  position: number;
}

export interface VmInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  url: string;
  phoneNumber: string;
  position: number;
}

export interface VmPatch {
  /** @minLength 1 */
  name?: string;
  /** @minLength 1 */
  url?: string;
  phoneNumber?: string;
  position?: number;
}

export interface ErrorResponse {
  error: string;
}