# lib/api-zod/src/generated/types/line.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ViP Dante VM Console API
 * OpenAPI spec version: 0.1.0
 */
import type { LineState } from './lineState';

export interface Line {
  id: string;
  state: LineState;
  label: string;
  updatedAt: Date;
}