# lib/api-zod/src/generated/types/lineInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ViP Dante VM Console API
 * OpenAPI spec version: 0.1.0
 */
import type { LineInputState } from './lineInputState';

export interface LineInput {
  state?: LineInputState;
  /** @maxLength 120 */
  label?: string;
}