# lib/api-zod/src/generated/types/vmPatch.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * ViP Dante VM Console API
 * OpenAPI spec version: 0.1.0
 */

export interface VmPatch {
  /** @minLength 1 */
  name?: string;
  /** @minLength 1 */
  url?: string;
  phoneNumber?: string;
  position?: number;
}